requirements.txt
#Flask==1.0.2
#Flask is python2 ...
flask
google-cloud-pubsub==0.39.1
gunicorn==19.9.0
google-cloud-datastore
google-cloud-storage
google-api-python-client
numpy<1.17 # To avoid gettting spammed warning https://github.com/tensorflow/tensorflow/issues/30427#issuecomment-515770145
tensorflow
requests